+++ src/parser/block/paragraph/mention.ts
@@ -1,11 +1,11 @@
 import { ParagraphParser } from '../../block';
 import { subsequence, some, block } from '../../../combinator';
-import { address } from './mention/address';
+import { anchor } from './mention/anchor';
 import { quote } from './mention/quote';
 
 export const mention: ParagraphParser.MentionParser = block(
   subsequence([
-    some(address),
+    some(anchor),
     quote,
   ]),
   false);