+++ src/parser/inline/autolink/anchor.test.ts
@@ -2,8 +2,8 @@ import { autolink } from '../autolink';
 import { some } from '../../../combinator';
 import { inspect } from '../../../debug.test';
 
-describe('Unit: parser/inline/autolink/address', () => {
-  describe('address', () => {
+describe('Unit: parser/inline/autolink/anchor', () => {
+  describe('anchor', () => {
     const parser = (source: string) => some(autolink)(source, {});
 
     it('invalid', () => {
@@ -14,8 +14,8 @@ describe('Unit: parser/inline/autolink/address', () => {
     });
 
     it('valid', () => {
-      assert.deepStrictEqual(inspect(parser('>0')), [['<a class="address" href="?res=0" rel="noopener">&gt;0</a>'], '']);
-      assert.deepStrictEqual(inspect(parser('>0aA')), [['<a class="address" href="?res=0aA" rel="noopener">&gt;0aA</a>'], '']);
+      assert.deepStrictEqual(inspect(parser('>0')), [['<a class="anchor" href="?res=0" rel="noopener">&gt;0</a>'], '']);
+      assert.deepStrictEqual(inspect(parser('>0aA')), [['<a class="anchor" href="?res=0aA" rel="noopener">&gt;0aA</a>'], '']);
     });
 
   });